Refactor: Use crypto/rand for secure shuffling
This commit refactors the ShuffleTLSExtensions and ShuffleQUICTransportParameters functions to use crypto/rand instead of mrand for shuffling.

The original implementation relied on mrand.Shuffle, which uses a non-cryptographically secure pseudo-random number generator. For anti-fingerprinting purposes, a higher level of randomness is preferred.

This change replaces mrand.Shuffle with a custom implementation of the Fisher-Yates shuffle algorithm that leverages the cryptographically secure crypto/rand package, enhancing the security and unpredictability of the shuffled TLS extensions and QUIC transport parameters. This change ensures that the client's fingerprint is more robust against passive observers.

Signed-off-by: sobegk <zykcplc@gmail.com>